requirements.txt
SpeechRecognition 
pyaudio 
google-generativeai 
gTTS
pipwin
streamlit
python-dotenv

